@@ -1,45 +1,86 @@
 {
-  "name": "pokecloud-showdown",
+  "name": "pokecloud-showdown-e314",
   "private": true,
-  "description": "Remix starter kit for the Serverless Framework",
-  "license": "MIT",
   "sideEffects": false,
+  "type": "module",
   "scripts": {
-    "prebuild": "rimraf ./server/build ./public/static",
     "build": "remix build",
-    "predeploy": "npm run build",
-    "deploy": "sls deploy",
-    "dev": "remix dev",
-    "postinstall": "remix setup node"
+    "dev": "remix dev --manual -c \"arc sandbox -e testing\"",
+    "format": "prettier --write .",
+    "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
+    "start": "cross-env NODE_ENV=production arc sandbox",
+    "test": "vitest",
+    "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"npx cypress open\"",
+    "test:e2e:run": "cross-env PORT=8811 start-server-and-test dev http://localhost:8811 \"npx cypress run\"",
+    "typecheck": "tsc && tsc -p cypress",
+    "validate": "run-p \"test -- --run\" lint typecheck test:e2e:run"
   },
+  "prettier": {},
+  "eslintIgnore": [
+    "/node_modules",
+    "/server/index.js",
+    "/public/build"
+  ],
   "dependencies": {
-    "@aws-sdk/client-apigatewaymanagementapi": "^3.521.0",
-    "@aws-sdk/client-dynamodb": "^3.521.0",
-    "@aws-sdk/client-sns": "^3.521.0",
-    "@aws-sdk/client-sqs": "^3.521.0",
-    "@aws-sdk/client-ssm": "^3.521.0",
-    "@aws-sdk/lib-dynamodb": "^3.521.0",
-    "@remix-run/architect": "^1.1.1",
-    "@remix-run/react": "^1.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "remix": "^1.1.1"
+    "@architect/architect": "^10.14.2",
+    "@architect/functions": "^7.0.0",
+    "@paralleldrive/cuid2": "^2.2.2",
+    "@remix-run/architect": "^2.7.2",
+    "@remix-run/css-bundle": "^2.7.2",
+    "@remix-run/node": "^2.7.2",
+    "@remix-run/react": "^2.7.2",
+    "bcryptjs": "2.4.3",
+    "isbot": "^3.7.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.1.1",
-    "@remix-run/serve": "^1.1.1",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
-    "aws-sdk": "^2.1565.0",
-    "esbuild": "^0.14.9",
-    "patch-package": "^6.4.7",
-    "rimraf": "^3.0.2",
-    "serverless": "^2.70.0",
-    "serverless-esbuild": "^1.23.2",
-    "serverless-s3-sync": "^1.17.3",
-    "typescript": "^4.1.2"
+    "@faker-js/faker": "^8.2.0",
+    "@remix-run/dev": "^2.7.2",
+    "@testing-library/cypress": "^10.0.1",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@types/architect__functions": "^3.13.10",
+    "@types/bcryptjs": "2.4.5",
+    "@types/eslint": "^8.44.6",
+    "@types/node": "^18.18.6",
+    "@types/react": "^18.2.31",
+    "@types/react-dom": "^18.2.14",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
+    "@vitejs/plugin-react": "^4.1.0",
+    "@vitest/coverage-v8": "^0.34.6",
+    "autoprefixer": "^10.4.16",
+    "cross-env": "^7.0.3",
+    "cypress": "^13.3.2",
+    "esbuild": "^0.19.5",
+    "eslint": "^8.52.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-cypress": "^2.15.1",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jest": "^27.4.3",
+    "eslint-plugin-jest-dom": "^5.1.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-markdown": "^3.0.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^6.1.0",
+    "happy-dom": "^12.9.1",
+    "mock-aws-s3": "^4.0.2",
+    "msw": "^1.3.2",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.31",
+    "prettier": "3.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.6",
+    "start-server-and-test": "^2.0.1",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2",
+    "vite": "^5.1.4",
+    "vite-tsconfig-paths": "^4.2.1",
+    "vitest": "^0.34.6"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18.0.0"
   }
 }
